+++ packages/admin-ui/ui/src/components/organisms/analytics-config-form/index.tsx
@@ -3,6 +3,8 @@ import { useEffect } from "react"
 import { Controller, useWatch } from "react-hook-form"
 import { NestedForm } from "../../../utils/nested-form"
 import Switch from "../../atoms/switch"
+import InfoIcon from "../../fundamentals/icons/info-icon"
+import Tooltip from "../../atoms/tooltip"
 
 export type AnalyticsConfigFormType = {
   anonymize: boolean
@@ -11,9 +13,10 @@ export type AnalyticsConfigFormType = {
 
 type Props = {
   form: NestedForm<AnalyticsConfigFormType>
+  compact?: boolean
 }
 
-const AnalyticsConfigForm = ({ form }: Props) => {
+const AnalyticsConfigForm = ({ form, compact }: Props) => {
   const { control, setValue, path } = form
 
   const watchOptOut = useWatch({
@@ -31,17 +34,33 @@ const AnalyticsConfigForm = ({ form }: Props) => {
   return (
     <div className="gap-y-xlarge flex flex-col">
       <div
-        className={clsx("flex items-start transition-opacity", {
+        className={clsx("flex items-center gap-3 transition-opacity", {
           "opacity-50": watchOptOut,
         })}
       >
         <div className="gap-y-2xsmall flex flex-1 flex-col">
-          <h2 className="inter-base-semibold">Anonymize my usage data</h2>
-          <p className="inter-base-regular text-grey-50">
-            You can choose to anonymize your usage data. If this option is
-            selected, we will not collect your personal information, such as
-            your name and email address.
-          </p>
+          <div className="flex items-center">
+            <h2 className="inter-base-semibold mr-2">
+              Anonymize my usage data{" "}
+            </h2>
+            {compact && (
+              <Tooltip
+                content="You can choose to anonymize your usage data. If this option is
+              selected, we will not collect your personal information, such as
+              your name and email address."
+                side="top"
+              >
+                <InfoIcon size="18px" color={"#889096"} />
+              </Tooltip>
+            )}
+          </div>
+          {!compact && (
+            <p className="inter-base-regular text-grey-50">
+              You can choose to anonymize your usage data. If this option is
+              selected, we will not collect your personal information, such as
+              your name and email address.
+            </p>
+          )}
         </div>
         <Controller
           name={path("anonymize")}
@@ -57,14 +76,26 @@ const AnalyticsConfigForm = ({ form }: Props) => {
           }}
         />
       </div>
-      <div className="flex items-start">
+      <div className="flex items-center gap-3">
         <div className="gap-y-2xsmall flex flex-1 flex-col">
-          <h2 className="inter-base-semibold">
-            Opt out of sharing my usage data
-          </h2>
-          <p className="inter-base-regular text-grey-50">
-            You can always opt out of sharing your usage data at any time.
-          </p>
+          <div className="flex items-center">
+            <h2 className="inter-base-semibold mr-2">
+              Opt out of sharing my usage data
+            </h2>
+            {compact && (
+              <Tooltip
+                content="You can always opt out of sharing your usage data at any time."
+                side="top"
+              >
+                <InfoIcon size="18px" color={"#889096"} />
+              </Tooltip>
+            )}
+          </div>
+          {!compact && (
+            <p className="inter-base-regular text-grey-50">
+              You can always opt out of sharing your usage data at any time.
+            </p>
+          )}
         </div>
         <Controller
           name={path("opt_out")}