+++ packages/admin-ui/ui/src/domain/gift-cards/manage/index.tsx
@@ -3,12 +3,14 @@ import { useAdminProducts } from "medusa-react"
 import { useNavigate } from "react-router-dom"
 import BackButton from "../../../components/atoms/back-button"
 import Spinner from "../../../components/atoms/spinner"
+import WidgetContainer from "../../../components/extensions/widget-container"
 import GiftCardDenominationsSection from "../../../components/organisms/gift-card-denominations-section"
 import ProductAttributesSection from "../../../components/organisms/product-attributes-section"
 import ProductGeneralSection from "../../../components/organisms/product-general-section"
 import ProductMediaSection from "../../../components/organisms/product-media-section"
 import ProductRawSection from "../../../components/organisms/product-raw-section"
 import ProductThumbnailSection from "../../../components/organisms/product-thumbnail-section"
+import { useWidgets } from "../../../providers/widget-provider"
 import { getErrorStatus } from "../../../utils/get-error-status"
 
 const Manage = () => {
@@ -24,6 +26,8 @@ const Manage = () => {
   )
 
   const giftCard = products?.[0] as Product | undefined
+
+  const { getWidgets } = useWidgets()
 
   if (!giftCard) {
     return (
@@ -57,9 +61,34 @@ const Manage = () => {
       />
       <div className="gap-x-base grid grid-cols-12">
         <div className="gap-y-xsmall col-span-8 flex flex-col">
+          {getWidgets("gift_card.details.before").map((w, i) => {
+            return (
+              <WidgetContainer
+                key={i}
+                widget={w}
+                injectionZone={"gift_card.details.before"}
+                entity={giftCard}
+              />
+            )
+          })}
+
           <ProductGeneralSection product={giftCard} />
+
           <GiftCardDenominationsSection giftCard={giftCard} />
+
           <ProductAttributesSection product={giftCard} />
+
+          {getWidgets("gift_card.details.after").map((w, i) => {
+            return (
+              <WidgetContainer
+                key={i}
+                widget={w}
+                injectionZone={"gift_card.details.after"}
+                entity={giftCard}
+              />
+            )
+          })}
+
           <ProductRawSection product={giftCard} />
         </div>
         <div className="gap-y-xsmall col-span-4 flex flex-col">