+++ packages/admin/src/lib/index.ts
@@ -1,0 +1,3 @@
+import { withCustomWebpackConfig } from "@medusajs/admin-ui"
+export * from "../types"
+export { withCustomWebpackConfig }